refactor: rename and relocate classes
