+++ src/main/java/com/github/spotbugs/snom/internal/SpotBugsTextReport.java
@@ -13,13 +13,14 @@
  */
 package com.github.spotbugs.snom.internal;
 
+import com.github.spotbugs.snom.SpotBugsReport;
 import com.github.spotbugs.snom.SpotBugsTask;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.io.File;
 import java.util.Optional;
 import org.gradle.api.model.ObjectFactory;
 
-public class SpotBugsTextReport extends AbstractSingleFileReport {
+public class SpotBugsTextReport extends SpotBugsReport {
   public SpotBugsTextReport(ObjectFactory objects, SpotBugsTask task) {
     super(objects, task);
     // the default reportsDir is "$buildDir/reports/spotbugs/${taskName}/spotbugs.txt"
